rename ScoreAnnou(,n)cing.tsx: fix minor issues
Issue: https://github.com/sopra-fs22-group-36/screw-your-neighbor-server/issues/8
