+++ src/components/ui/ScoreAnnouncing.tsx
